Attempt to parse all non-JSON files
This allows proxyquireify to be combined with compiled/transpiled code, e.g.
ES6 and CoffeeScript

Breaking: transforms must now be correctly ordered so that all non-JSON files
that reach proxyquireify's transform are valid JS.

".json" is a special case because Browserify transforms it to valid JS
outside of the normal transform pipeline.

Closes #24